+++ src/locks/mod.rs
@@ -1,4 +1,4 @@
 //! Provides simplified lock types for sharing data between threads
 
-pub mod sync;
+pub mod std;
 pub mod tokio;
